// src/lib/api/schemas/cycleTimeKpi.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */

export interface CycleTimeKpi {
  /** @nullable */
  average_seconds?: number | null;
  /** @nullable */
  median_seconds?: number | null;
  /** @nullable */
  p75_seconds?: number | null;
  /** @nullable */
  p90_seconds?: number | null;
  sample_size: number;
}