// src/lib/api/schemas/salesKpis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { BrandSalesKpi } from './brandSalesKpi.ts';
import type { CustomerConcentration } from './customerConcentration.ts';
import type { CustomerKpis } from './customerKpis.ts';
import type { PaymentTermSalesKpi } from './paymentTermSalesKpi.ts';
import type { VariantSalesKpi } from './variantSalesKpi.ts';

export interface SalesKpis {
  average_lines_per_order: number;
  average_net_unit_price: number;
  average_order_value: number;
  average_units_per_order: number;
  brands: BrandSalesKpi[];
  customer_concentration: CustomerConcentration;
  customers: CustomerKpis;
  discount_rate: number;
  from: string;
  line_count: number;
  order_count: number;
  payment_terms: PaymentTermSalesKpi[];
  sales_total: number;
  tax_rate: number;
  timezone: string;
  to: string;
  units_sold: number;
  variants: VariantSalesKpi[];
}